refactor(bezier-icons): make utils directory for utility function
