+++ packages/bezier-icons/scripts/generate-icon-files.js
@@ -1,6 +1,6 @@
 const path = require('path')
 
-const { generateIconFiles } = require('./generateIconFilesFromJson')
+const { generateIconFiles } = require('./utils/generateIconFilesFromJson')
 
 const bezierIconsDirectory = path.resolve(__dirname, '..')
 const iconsJson = path.resolve(bezierIconsDirectory, 'icons.json')
