+++ packages/bezier-icons/scripts/utils/generateIconFilesFromJson.spec.ts
