+++ packages/bezier-icons/scripts/utils/generateIconFilesFromJson.ts
